requirements_pi.txt
Adafruit-Blinka==6.4.2
adafruit-circuitpython-74hc595==1.2.4
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-charlcd==3.3.9
adafruit-circuitpython-mcp230xx==2.4.5
Adafruit-PlatformDetect==3.5.0
Adafruit-PureIO==1.1.8
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
colorzero==1.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dlib==19.20.0
entrypoints==0.3
gpiozero==1.5.1
importlib-metadata==1.6.0
imutils==0.5.3
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.20.1
opencv-contrib-python==4.1.1.26
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
picamera==1.13
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyftdi==0.52.9
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.16.0
pyserial==3.5
python-dateutil==2.8.1
pytorch-lamb==1.0.0
pytz==2020.1
pyusb==1.1.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
recordtype==1.3
rpi-ws281x==4.2.6
RPi.GPIO==0.7.0
scikit-learn==0.23.1
scipy==1.5.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
spidev==3.5
sysv-ipc==1.1.0
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0